requirements.txt
requests==2.31.0
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
google-cloud-bigquery==3.19.0
